feat: notifications, removing pusher channel
