+++ pages/home/index.html
@@ -26,11 +26,7 @@
   <!-- Feature JS -->
   <script type="module" src="./script.js" defer></script>
 
-  <!-- Notifications JS -->
-  <script src="https://js.pusher.com/beams/1.0/push-notifications-cdn.js"></script>
-  <script src="https://js.pusher.com/8.2.0/pusher.min.js"></script>
-  <script src="../../assets/helpers/push-notifications.js" type="module"></script>
-  <!-- End Notifications JS -->
+
 
   <title>Outsafe BC</title>
 </head>
@@ -38,6 +34,11 @@
 <body id="home-body">
   <main id="root"></main>
   <noscript>You need to enable JavaScript to run this app.</noscript>
+
+  <!-- Notifications JS -->
+  <script src="https://js.pusher.com/beams/1.0/push-notifications-cdn.js"></script>
+  <script src="https://js.pusher.com/8.2.0/pusher.min.js"></script>
+  <!-- End Notifications JS -->
 </body>
 
 </html>
