+++ pages/login/index.html
@@ -34,7 +34,6 @@
 	<!-- Notifications JS -->
 	<script src="https://js.pusher.com/beams/1.0/push-notifications-cdn.js"></script>
 	<script src="https://js.pusher.com/8.2.0/pusher.min.js"></script>
-	<script src="../../assets/helpers/push-notifications.js" type="module" defer></script>
 	<!-- End Notifications JS -->
 
 	<title>Login</title>
